+++ pkg/ccr/meta.go
@@ -1164,3 +1164,11 @@ func (m *Meta) ClearTable(dbName string, tableName string) {
 
 	delete(m.TableName2IdMap, tableName)
 }
+
+func (m *Meta) IsPartitionDropped(partitionId int64) bool {
+	panic("IsPartitionDropped is not supported, please use ThriftMeta instead")
+}
+
+func (m *Meta) IsTableDropped(partitionId int64) bool {
+	panic("IsTableDropped is not supported, please use ThriftMeta instead")
+}